requirements.txt
gradio>=4.0.0
requests>=2.28.0
pandas>=1.5.0
plotly>=5.15.0
opencc-python-reimplemented>=0.1.7
